Troquei as estilizações para componentes styles.js separados
