+++ modulo3/projeto-astromatch/src/Components/Matches/MatchesFooter.js
@@ -1,29 +1,18 @@
 import React from 'react'
-import styled from 'styled-components';
+import {FooterContainer, Botao} from './styles'
+import clearIcon from './img/clear.png'
 
-const Footer = styled.div`
-background-color: white;
-display: flex;
-justify-content: space-around;
-align-items: center;
-border: 1px solid #02afcc;
-border-bottom-left-radius: 15px;
-border-bottom-right-radius: 15px;
-padding: 0 20px;
-width: 20vw;
-height: 8vh;
-margin: 10px auto;
-position: relative;
-h2{
-    color:#02afcc;
-}
-`
+function MatchesFooter (props) {
+    const clearMatches = () => {
+        props.clear();
+        props.getMatches()
+      }
 
-function MatchesFooter () {
     return (    
-    <Footer>
+    <FooterContainer>
         <h2>Your Matches!</h2>
-    </Footer>)
+        <Botao onClick={clearMatches}><img src={clearIcon}/></Botao>
+    </FooterContainer>)
 }
 
 export default MatchesFooter;
